moved onReleaseCB logic to helpers.js
